feat: add validators to person Controller

+++ src/controllers/personController.js
@@ -3,7 +3,9 @@ const ValidationContract = require('../util/validators');
 
 exports.createPerson = async (req, res) => {
   let validators = new ValidationContract();
-  //validators.isRequired(req.body.payroll, 'Payroll is required');
+  validators.isRequired(req.body.name, 'name is required');
+  validators.isRequired(req.body.cpf, 'cpf is required');
+  validators.isRequired(req.body.address, 'address is required');
 
   try {
     if (validators.isValid()) {
@@ -25,7 +27,7 @@ exports.findAllPersons = async (req, res) => {
   try {
     const persons = await personRepository.findAllPersons();
     if (persons == null || persons.length == 0) {
-      res.status(204).send('No Persons found');
+      res.status(204).send();
     } else {
       res.status(200).send(persons);
     }
@@ -37,27 +39,81 @@ exports.findAllPersons = async (req, res) => {
 };
 
 exports.findPersonById = async (req, res) => {
+  let validators = new ValidationContract();
   const personId = req.params.id;
-  if (personId == null) {
-    res.status(400).send('PersonId is required');
-  }
-  const person = personRepository.findPersonById(personId);
+  validators.isRequired(personId, 'personId is required');
+  validators.isObjectIdValid(
+    personId,
+    `personId: "${personId}" is not a ObjectId valid.`,
+  );
 
-  if (!person) {
-    res.status(404).send();
+  try {
+    if (validators.isValid()) {
+      const person = await personRepository.findPersonById(personId);
+      if (!person) {
+        res.status(204).send();
+        return;
+      }
+      res.status(200).send(person);
+    } else {
+      res.status(400).send({
+        errors: validators.getErrors(),
+      });
+    }
+  } catch (error) {
+    res.status(500).send({
+      message: 'Server error.',
+    });
   }
-
-  res.status(200).send(person);
 };
 
 exports.updatePersonById = async (req, res) => {
+  let validators = new ValidationContract();
   const personId = req.params.id;
-  await personRepository.updatePersonById(personId, req.body);
-  res.status(200).send('Person updated', req.body);
+  validators.isRequired(personId, 'personId is required');
+  validators.isObjectIdValid(
+    personId,
+    `personId: "${personId}" is not a ObjectId valid.`,
+  );
+
+  try {
+    if (validators.isValid()) {
+      await personRepository.updatePersonById(personId, req.body);
+      res.status(200).send('Person updated');
+    } else {
+      res.status(400).send({
+        errors: validators.getErrors(),
+      });
+    }
+  } catch (error) {
+    res.status(500).send({
+      message: 'Server error.',
+    });
+  }
 };
 
 exports.deletePersonById = async (req, res) => {
+  let validators = new ValidationContract();
   const personId = req.params.id;
-  await personRepository.deletePersonById(personId);
-  res.status(204).send('Person deleted', req.body);
+  validators.isRequired(personId, 'personId is required');
+  validators.isObjectIdValid(
+    personId,
+    `personId: "${personId}" is not a ObjectId valid.`,
+  );
+
+  try {
+    if (validators.isValid()) {
+      await personRepository.deletePersonById(personId);
+      res.status(204).send();
+    } else {
+      res.status(400).send({
+        errors: validators.getErrors(),
+      });
+    }
+  } catch (error) {
+    console.log(error);
+    res.status(500).send({
+      message: 'Server error.',
+    });
+  }
 };
